refactor(ui): move the equipment type metadata into its own module
Clears the mechanical half of #796.

`react-refresh/only-export-components` exists to keep fast refresh
working: a module exporting both a component and something else loses it,
so an edit there reloads the page instead of preserving state.
EQUIPMENT_TYPE_CATEGORIES, TYPE_ICONS and TYPE_LABELS sat next to the
components that read them. Version 0.4 let them through under
allowConstantExport, 0.5 does not, and rewriting them as inline
`export const` does not satisfy it either (measured before choosing this
route). Moving them is what the rule was asking for.

They now live in equipment-type-meta.tsx, imported by DeviceSelector,
EquipmentCard, useEquipmentState and EquipmentDetailPage. The rule is
back to the preset's error, so nothing can drift back.

The other half of #796 is a decision, not a refactor:
react-hooks/set-state-in-effect is switched off rather than left at warn.
Its twenty sites are all on-mount fetches, `useEffect(() => { load(); },
[])`, and the rule's concern is the extra render pass that causes. It
does not bite here; the pages load correctly, verified across ten routes
on the release candidate. Twenty permanent warnings on every lint run is
the volume at which people stop reading the output, and rewriting the
effect timing of twenty working screens for an opinionated rule trades a
real regression risk for no visible gain. #796 keeps the list, so if a
screen ever does flicker or loop the places to look are written down.

Also removes nine now-dead eslint-disable directives for that rule.

Note for anyone repeating this: ui/ is not covered by the repo's
format:check, which only globs the backend. Running prettier across
ui/src rewrites 184 files. Only the new module was formatted here.
