+++ ui/eslint.config.js
@@ -20,20 +20,21 @@ export default defineConfig([
       globals: globals.browser,
     },
     rules: {
-      // Demoted to warn, not switched off, when /ui moved to eslint 10 with
-      // eslint-plugin-react-hooks 7.1 and react-refresh 0.5. Both rules
-      // tightened and between them report 23 real findings across 20 files.
-      // They are genuine, and none is a live bug: the set-state-in-effect hits
-      // are on-mount `useEffect(() => { load(); }, [])` fetches, and the
-      // only-export-components hits are three constant maps exported next to
-      // the component that owns them.
+      // react-hooks/set-state-in-effect tightened in eslint-plugin-react-hooks
+      // 7.1 and reports 20 sites, every one of them an on-mount fetch of the
+      // shape `useEffect(() => { load(); }, [])`: open a page, go and get its
+      // data, store it. The rule's concern is the extra render pass that
+      // causes, which can show as a flicker or a loop. It does not here: the
+      // pages load correctly, verified across ten routes on the v1.61.0
+      // release candidate.
       //
-      // Fixing them means touching effect timing in 20 components, which is
-      // behaviour, not tooling. That work is tracked and reviewed on its own
-      // rather than smuggled into a dependency bump. See the tracking issue
-      // tracked in #796.
-      "react-hooks/set-state-in-effect": "warn",
-      "react-refresh/only-export-components": "warn",
+      // Off rather than left at warn, on purpose. Twenty permanent warnings on
+      // every lint run is the volume at which people stop reading the output,
+      // and rewriting the effect timing of twenty working screens to satisfy an
+      // opinionated rule trades a real regression risk for no visible gain.
+      // #796 keeps the list of the twenty, so if a screen ever does flicker or
+      // loop, the places to look are already written down.
+      "react-hooks/set-state-in-effect": "off",
     },
   },
 ]);
