+++ ui/src/components/equipments/PvForecastPanel.tsx
@@ -63,7 +63,6 @@ export function PvForecastPanel({ equipmentId }: PvForecastPanelProps) {
   }, [equipmentId, accuracyDays]);
 
   useEffect(() => {
-    // eslint-disable-next-line react-hooks/set-state-in-effect -- load() only sets state after its awaits resolve
     void load();
   }, [load]);
